[models] clean up naming
When returning data for a user, serialize the user's serverdata. This is
why 'user' foreignkey in UserServerData is called serverdata.
When returning data for a server, serialize the server's userdata. THis
is why 'server' foreignkey is UserServerData is called userdata.
Update other names to be consistent with this.
Update serializer naming to be very explicit.
Add UserServerData serializer.